project's new folder structure
tried to revamp existing folder structure with this more context focused approach, and also hoping to ommit previously existed cyclick module dependency for category selectrion and ending screen call.
