+++ src/all-components/question-generator/randomizer.js
@@ -1,4 +1,4 @@
-import data from './data.json';
+import data from '../../data.json';
 
 /**
  * @typedef {Object} QuestionIterator
