+++ src/all-components/starting-screen/categorySelection.js
@@ -1,0 +1,12 @@
+import { $ } from "../../all-utils/for-dom-calls";
+import { chosenCategory } from "../../all-utils/for-starting-screen";
+
+
+export let commenceCategorySelection = () => {
+    $('.categories-container').addEventListener('click', handleCategorySelection);
+};
+
+let handleCategorySelection = (evt) => {
+    let category = evt.target.parentNode.classList[0];
+    chosenCategory(category);
+};
