+++ src/all-components/starting-screen/showLayout.js
@@ -1,4 +1,4 @@
-import { createStartingScreenLayout } from '../all-layouts/startingScreen';
+import { createStartingScreenLayout } from "../../all-layouts/startingScreen";
 
 export let addStartingScreenToDOM = () => {
     let startingScreen = createStartingScreenLayout();
