+++ src/all-utils/for-dom-calls/index.js
@@ -1,0 +1,4 @@
+let $ = selector => document.querySelector(selector);
+let clearScreen = () => document.body.innerHTML = '';
+
+export {$, clearScreen}
