--- src/startup-components/mainMenu.js
@@ -1,14 +1,0 @@
-import { MainMenu } from '../all-layouts/MainMenu';
-import { addStartingScreenToDOM } from './startingScreen';
-import { commenceCategorySelection } from './userInteraction';
-
-export const showMenu = () => {
-    document.body.innerHTML = '';
-    document.body.append(MainMenu());
-
-    document.querySelector('#playBtn').addEventListener('click', () => {
-        document.body.innerHTML = '';
-        addStartingScreenToDOM();
-        commenceCategorySelection();
-    });
-};
